[update] .config/nvim install.sh .vimrc

+++ install.sh
@@ -5,25 +5,28 @@
 set -ue
 
 SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE:-$0}")" && pwd -P)
+BACKUP_DATE=$(date "+%Y%m%d-%H%M%S")
 
 create_symlink_dotfiles() {
-    command echo "backup old dotfiles..."
-    if [ ! -d "$HOME/.dotbackup" ]; then
-        command echo "$HOME/.dotbackup not found. Auto Make it"
-        command mkdir -v "$HOME/.dotbackup"
+    command echo "$BACKUP_DATE backup old dotfiles..."
+    if [ ! -d "$HOME/.dotbackup/$BACKUP_DATE" ]; then
+        command echo "$HOME/.dotbackup/$BACKUP_DATE not found. Auto Make it"
+        command mkdir -pv "$HOME/.dotbackup/$BACKUP_DATE"
     fi
 
     local dot_dir="${SCRIPT_DIR}"
+    local backup_dir="$HOME/.dotbackup/$BACKUP_DATE"
     if [[ "$HOME" != "$dot_dir" ]]; then
         for file in $dot_dir/.??*; do
             [[ `basename $file` == ".git" ]] && continue
             [[ `basename $file` == ".gitignore" ]] && continue
             [[ `basename $file` == ".config" ]] && continue
+            [[ `basename $file` == "scripts" ]] && continue
             if [[ -L "$HOME/`basename $file`" ]]; then
                 command rm -f "$HOME/`basename $file`"
             fi
             if [[ -e "$HOME/`basename $file`" && ! -L "$HOME/`basename $file`" ]]; then
-                command mv "$HOME/`basename $file`" "$HOME/.dotbackup"
+                command mv "$HOME/`basename $file`" "$backup_dir"
             fi
             command ln -snfv $file $HOME
         done
@@ -34,9 +37,9 @@ create_symlink_dotfiles() {
 
 create_symlink_config() {
     command echo "backup old dotconfig..."
-    if [ ! -d "$HOME/.dotbackup/.config" ]; then
-        command echo "$HOME/.dotbackup/.config not found. Auto Make it"
-        command mkdir -pv "$HOME/.dotbackup/.config"
+    if [ ! -d "$HOME/.dotbackup/$BACKUP_DATE/.config" ]; then
+        command echo "$HOME/.dotbackup/$BACKUP_DATE/.config not found. Auto Make it"
+        command mkdir -pv "$HOME/.dotbackup/$BACKUP_DATE/.config"
     fi
 
     if [ ! -d "$HOME/.config" ]; then
@@ -46,15 +49,17 @@ create_symlink_config() {
 
     local config_dir="${SCRIPT_DIR}/.config"
     local home_config="$HOME/.config"
+    local backup_config_dir="$HOME/.dotbackup/$BACKUP_DATE/.config"
     if [[ "$HOME" != "$config_dir" ]]; then
         for file in $config_dir/??*; do
             if [ -d $file ]; then
                 [[ `basename $file` = ".git" ]] && continue
+                [[ `basename $file` = "wezterm" ]] && continue
                 if [[ -L "$home_config/`basename $file`" ]]; then
                     command rm -f "$home_config/`basename $file`"
                 fi
                 if [[ -e "$home_config/`basename $file`" && ! -L "$home_config/`basename $file`" ]]; then
-                    command mv "$home_config/`basename $file`" "$HOME/.dotbackup/.config"
+                    command mv "$home_config/`basename $file`" "$backup_config_dir"
                 fi
                 command ln -snfv $file $home_config
             fi
